Move geojson to feature collection class
