--- sql2gee/ImageCollection.py
@@ -1,8 +1,0 @@
-class ImageCollection(Collection):
-  """docstring for ImageCollection"""
-  def __init__(self, arg):
-    super(ImageCollection, self).__init__()
-    self.arg = arg
-    
-  def response():
-    pass
