+++ sql2gee/image_collection.py
@@ -1,0 +1,11 @@
+from collection import Collection
+
+class ImageCollection(object):
+  """docstring for ImageCollection"""
+  def __init__(self, arg):
+    super(ImageCollection, self).__init__()
+    self.arg = arg
+    
+  def response():
+    pass
+    # ImageCollection.<filters>.<functions>.<sorts>.<imageReducers>.limit(n).getInfo()
